+++ a_sync/__init__.py
@@ -1,18 +1,17 @@
 
 from a_sync import aliases, exceptions, iter, task
-from a_sync.base import ASyncGenericBase
-from a_sync.decorator import a_sync
+from a_sync.a_sync import ASyncGenericBase, ASyncGenericSingleton, a_sync
+from a_sync.a_sync.modifiers.semaphores import apply_semaphore
+from a_sync.a_sync.property import ASyncCachedPropertyDescriptor, ASyncPropertyDescriptor, cached_property, property
+from a_sync.asyncio import as_completed, create_task, gather
 from a_sync.future import ASyncFuture, future  # type: ignore [attr-defined]
 from a_sync.iter import ASyncFilter as filter
 from a_sync.iter import ASyncSorter as sorted
 from a_sync.iter import ASyncIterable, ASyncIterator
-from a_sync.modifiers.semaphores import apply_semaphore
 from a_sync.primitives import *
-from a_sync.property import ASyncCachedPropertyDescriptor, ASyncPropertyDescriptor, cached_property, property
-from a_sync.singleton import ASyncGenericSingleton
 from a_sync.task import TaskMapping as map
 from a_sync.task import TaskMapping, create_task
-from a_sync.utils import all, any, as_completed, as_yielded, gather
+from a_sync.utils import all, any, as_yielded
 
 # I alias the aliases for your convenience.
 # I prefer "aka" but its meaning is not intuitive when reading code so I created both aliases for you to choose from.
@@ -38,6 +37,7 @@ __all__ = [
     "as_completed",
 
     # functions
+    "a_sync",
     "all",
     "any",
     "as_yielded",
