+++ a_sync/a_sync/_descriptor.py
@@ -1,9 +1,9 @@
 
 import functools
 
-from a_sync import decorator
 from a_sync._typing import *
-from a_sync.modified import ASyncFunction, ModifiedMixin, ModifierManager
+from a_sync.a_sync import decorator
+from a_sync.a_sync.function import ASyncFunction, ModifiedMixin, ModifierManager
 
 if TYPE_CHECKING:
     from a_sync import TaskMapping
