+++ a_sync/a_sync/abstract.py
@@ -3,9 +3,10 @@ import abc
 import functools
 import logging
 
-from a_sync import _flags, _kwargs, exceptions, modifiers
-from a_sync._meta import ASyncMeta
+from a_sync import exceptions
 from a_sync._typing import *
+from a_sync.a_sync import _flags, _kwargs, modifiers
+from a_sync.a_sync._meta import ASyncMeta
 from a_sync.exceptions import NoFlagsFound
 
 logger = logging.getLogger(__name__)
