+++ a_sync/a_sync/singleton.py
@@ -1,6 +1,6 @@
 
-from a_sync.base import ASyncGenericBase
-from a_sync._meta import ASyncSingletonMeta
+from a_sync.a_sync._meta import ASyncSingletonMeta
+from a_sync.a_sync.base import ASyncGenericBase
 
 class ASyncGenericSingleton(ASyncGenericBase, metaclass=ASyncSingletonMeta):
     """
