+++ a_sync/utils/__init__.py
@@ -2,16 +2,12 @@ import asyncio
 
 from a_sync.utils.iterators import (as_yielded, exhaust_iterator,
                                     exhaust_iterators)
-from a_sync.utils.as_completed import as_completed
-from a_sync.utils.gather import gather
 
 
 __all__ = [
     #"all",
     #"any",
-    "as_completed",
     "as_yielded",
-    "gather",
     "exhaust_iterator",
     "exhaust_iterators",
 ]
